+++ py2max/__init__.py
@@ -35,7 +35,7 @@ Example:
     >>> p.save()
 """
 
-__version__ = "0.3.3"
+__version__ = "0.3.4"
 
 from .core import Box, Patcher, Patchline
 from .exceptions import (